requirements.txt
PyMuPDF>=1.24.0
python-docx>=1.1.0
pyspellchecker>=0.8.0
spylls>=0.1.7
pyinstaller>=6.0.0
